Restructure controller files
